serverless functions/sentence_length/requirements.txt
Flask==1.0.2
requests
google-cloud-firestore